[MAZ-137] harden framebuffer cap against integer overflow (CodeRabbit)
Make the framebuffer cap trustworthy on pathological inputs, the whole point of
this change:

- gpu.go: compute fbSize in uint64 so a buggy hypervisor reporting absurd
  dimensions can't wrap uint32 and slip past the cap. Once within the (uint32)
  cap the value fits uint32. (Dropped CodeRabbit's redundant second overflow
  check -- unreachable once the cap check is 64-bit.)
- init.go: the resolution-fitting loop compared dispW*dispH in uint32, which can
  itself overflow before the cap loop runs; widen both sides to uint64.
  (CodeRabbit flagged gpu.go but missed this matching wrap.)
- main.go + kernel_config.go: framebuffer_max_mb is operator-supplied; reject a
  negative value or one large enough that *1MiB wraps uint32, rather than
  silently installing a bogus (tiny) cap. New constants.MaxFramebufferMB (4095)
  is the single source for the bound.

Both arches build clean; host tests pass.

Refs: MAZ-137

+++ kmazarin/device/virtio/gpu/init.go
@@ -50,7 +50,9 @@ func Init() bool {
 	// Cap to what fits under framebuffer_max_mb (with 2x height for scrolling):
 	// maxBytes / 4 bytes-per-pixel / 2 (resource height) pixels.
 	maxPixels := framebufferMaxBytes / 4 / 2
-	for dispW*dispH > maxPixels {
+	// 64-bit comparison: a hypervisor-reported dispW*dispH can itself overflow
+	// uint32 before the cap loop runs, so widen both sides.
+	for uint64(dispW)*uint64(dispH) > uint64(maxPixels) {
 		// Scale down by 75% until it fits.
 		dispW = dispW * 3 / 4
 		dispH = dispH * 3 / 4
